pyproject.toml
[project]
name = "speaksee"
version = "0.1.0"
description = "Speak \u2192 See: fully local speech-to-image desktop web app"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.128.5",
  "uvicorn[standard]==0.39.0",

  # STT
  "faster-whisper==1.2.1",
  "numpy==1.26.4",

  # Image gen
  "torch==2.8.0",
  "diffusers==0.36.0",
  "transformers==4.57.6",
  "accelerate==1.10.1",
  "safetensors==0.7.0",
  "huggingface-hub==0.36.2",
  "pillow==11.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.2",
]
tts = [
  "pyttsx3==2.99",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
